refactor: add db backend wrapper for future postgres
